+++ Tests/Editor/SerializationSnapshotTestAsset.cs
@@ -1,5 +1,6 @@
 using System;
 using UnityEngine;
+using UnityEngine.Localization;
 
 namespace VMFramework.Pipeline.Editor.Tests
 {
@@ -20,5 +21,8 @@ namespace VMFramework.Pipeline.Editor.Tests
         public Rect rect;
         public Gradient gradient = new();
         public AnimationCurve curve = new();
+        public LocalizedString localizedByName = new();
+        public LocalizedString localizedByGuid = new();
+        [NonSerialized] public int runtimeValue = 5;
     }
 }